Use int64 for fields that require it
Golang's int is 32 or 64 bits depending on the platform so int64 should be used explicitly to
ensure these fields don't overflow.

I did not put much thought into backwards compatibility.

+++ models/audio.go
@@ -9,6 +9,6 @@ type Audio struct {
 	Title        string     `json:"title,omitempty"`
 	FileName     string     `json:"file_name,omitempty"`
 	MimeType     string     `json:"mime_type,omitempty"`
-	FileSize     int        `json:"file_size,omitempty"`
+	FileSize     int64      `json:"file_size,omitempty"`
 	Thumb        *PhotoSize `json:"thumb,omitempty"`
 }
